Add files via upload
This is a first draft, for entertainment purposes only.

The results show a player will lose 42-44% of hands played, but might get an average profit of 1.29% if the betting system is followed.

Double-down winning ratios came out to about 3 to 2, and split hands came to about 8 to 7.  This is for a back test of several million hands.  Being able to track the number of 10s and face cards in the last 12 shown was necessary to improve the win ratio.  Without that, a player could expect to lose about 10%.